Add security annotations in services (ExternalService isn't finished) Code cleanup :   remove unused imports   add @Override annotations   use Generics collections   remove deprecated CvqPermissionExceptions   remove useless casts   use Java5 foreach instead of Iterators when relevant   remove useless empty alone constructors

+++ Plugins/PaymentProviders/Spplus/src/java/fr/capwebct/capdemat/plugins/paymentproviders/spplus/service/SpplusService.java
@@ -69,7 +69,7 @@ public class SpplusService implements IPaymentProviderService {
         parameters.append("FR")
             .append("978");
         
-        String formattedTotal = numberFormat.format((double) payment.getAmount().doubleValue()/100);
+        String formattedTotal = numberFormat.format(payment.getAmount().doubleValue()/100);
         urlParameters.append("&montant=").append(formattedTotal);
         parameters.append(formattedTotal);
